создал модули, разнес код
